WPI-PR/requirements.txt
beautifulsoup4>=4.9.3
lxml>=4.6.3
PyYAML>=5.4.1
